pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alibabacloud-oss-v2>=1.2.5",
    "chromadb>=1.5.7",
    "dashscope>=1.25.16",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.1",
    "fastmcp>=3.2.4",
    "ipywidgets>=8.1.8",
    "jq>=1.11.0",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.1",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=1.1.0",
    "langchain-tavily>=0.2.18",
    "langchain-text-splitters>=1.1.1",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langgraph-cli[inmem]>=0.4.24",
    "openai>=2.30.0",
    "pypdf>=6.10.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "tavily-python>=0.7.24",
    "zhipuai>=1.0.7",
]
